refactor(ofm_core/contracts): extract expiry cluster
Move the contract-expiry sweep out of contracts/mod.rs into contracts/expiry.rs:
contract_warning_stage, process_contract_expiries, the shared release_player_contract
that both expiry and manager-termination funnel through, remove_player_from_team_references
and contract_expired_message. Byte-faithful move via script; two public entry points
re-exported via `pub use`, release_player_contract promoted to pub(crate). The shared
ContractReleaseReason enum stays in mod.rs (both expiry and termination construct/consume
it) and is bumped to pub(crate) to match the re-exported release routine's signature.

387 ofm_core lib + 28 contracts integration tests pass; clippy clean in contracts/.
